requirements.txt
asgiref==3.8.1
Django==5.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-smart-selects==1.7.2
pillow==10.4.0
psycopg2-binary==2.9.9
pytils==0.4.1
sqlparse==0.5.1
typing_extensions==4.12.2

django-multiupload~=0.6.1
django-phonenumber-field~=8.0.0